+++ pymatting/laplacian/cf_laplacian.py
@@ -3,7 +3,7 @@ import scipy.sparse
 from pymatting_aot.aot import _cf_laplacian
 
 
-def cf_laplacian(image, epsilon=1e-7, radius=1):
+def cf_laplacian(image, epsilon=1e-7, radius=1, is_known=None):
     """
     This function implements the alpha estimator for closed-form alpha matting as proposed by :cite:`levin2007closed`.
 
@@ -17,6 +17,9 @@ def cf_laplacian(image, epsilon=1e-7, radius=1):
        Radius of local window size, defaults to :math:`1`, i.e. only adjacent pixels are considered.
        The size of the local window is given as :math:`(2 r + 1)^2`, where :math:`r` denotes         the radius. A larger radius might lead to violated color line constraints, but also
        favors further propagation of information within the image.
+    is_known: numpy.ndarray
+        Binary mask of pixels for which to compute the laplacian matrix.
+        Laplacian entries for known pixels will have undefined values.
 
     Returns
     -------
@@ -26,12 +29,17 @@ def cf_laplacian(image, epsilon=1e-7, radius=1):
     h, w, d = image.shape
     n = h * w
 
+    if is_known is None:
+        is_known = np.zeros((h, w), dtype=np.bool8)
+
+    is_known = is_known.reshape(h, w)
+
     # Data for matting laplacian in csr format
     indptr = np.zeros(n + 1, dtype=np.int64)
     indices = np.zeros(n * (4 * radius + 1) ** 2, dtype=np.int64)
     values = np.zeros((n, 4 * radius + 1, 4 * radius + 1), dtype=np.float64)
 
-    _cf_laplacian(image, epsilon, radius, values, indices, indptr)
+    _cf_laplacian(image, epsilon, radius, values, indices, indptr, is_known)
 
     L = scipy.sparse.csr_matrix((values.ravel(), indices, indptr), (n, n))