+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.1.3
+- Add optimization for `estimate_alpha_cf` which should reduce computation time if most pixels in the trimap are known.
+- Allow sloppy trimaps.
+
 ### 1.1.2
 
 - Recompile ahead-of-time-compiled modules if they are out of date.